--order service(25%) --

+++ src/main/java/com/josh/SpringEcom/dto/OrderItemResponse.java
@@ -1,8 +1,9 @@
 package com.josh.SpringEcom.dto;
 
+import java.math.BigDecimal;
+
 public record OrderItemResponse(
         String productName,
         int quantity,
-        BigDecimal totalPrice
-) {
-}
+        int totalPrice
+) {}
